requirements.txt
Django==4.0.6
environs[django]==9.5.0
dj-database-url==0.5.0
whitenoise==6.2.0
python-telegram-bot==13.13
Faker==13.15.1
Jinja2==3.1.2

# Don't instal locally
gunicorn==20.1.0
psycopg2==2.9.3
